+++ al-khaser/Anti VM/Generic.h
@@ -28,3 +28,7 @@ BOOL disk_size_getdiskfreespace();
 BOOL accelerated_sleep();
 BOOL cpuid_is_hypervisor();
 BOOL cpuid_hypervisor_vendor();
+BOOL serial_number_bios_wmi();
+BOOL model_computer_system_wmi();
+BOOL manufacturer_computer_system_wmi();
+BOOL current_temperature_acpi_wmi();