+++ al-khaser/Anti VM/VirtualPC.cpp
@@ -22,3 +22,26 @@ VOID virtual_pc_process()
 			print_results(FALSE, msg);
 	}
 }
+
+
+
+VOID virtual_pc_reg_keys()
+{
+	/* Array of strings of blacklisted registry keys */
+	TCHAR* szKeys[] = {
+		_T("SOFTWARE\\Microsoft\\Virtual Machine\\Guest\\Parameters"),
+	};
+
+	WORD dwlength = sizeof(szKeys) / sizeof(szKeys[0]);
+
+	/* Check one by one */
+	for (int i = 0; i < dwlength; i++)
+	{
+		TCHAR msg[256] = _T("");
+		_stprintf_s(msg, sizeof(msg) / sizeof(TCHAR), _T("Checking reg key %s: "), szKeys[i]);
+		if (Is_RegKeyExists(HKEY_LOCAL_MACHINE, szKeys[i]))
+			print_results(TRUE, msg);
+		else
+			print_results(FALSE, msg);
+	}
+}