+++ al-khaser/Anti VM/VirtualPC.h
@@ -3,3 +3,4 @@
 
 
 VOID virtual_pc_process();
+VOID virtual_pc_reg_keys();