+++ al-khaser/Anti VM/Xen.h
@@ -3,3 +3,4 @@
 
 
 VOID xen_process();
+BOOL xen_check_mac();